# services/tweet-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  http.CreateTweetRequest:
    description: Request body for creating a tweet
    properties:
      content:
        example: Hello, this is my first tweet!
        type: string
    required:
    - content
    type: object
  http.ErrorResponse:
    description: Error response
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  http.Tweet:
    description: Tweet information
    properties:
      content:
        example: Hello, this is my first tweet!
        type: string
      created_at:
        example: "2024-06-07T22:04:25Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        example: "2024-06-07T22:04:25Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the tweet service API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tweet Service API
  version: "1.0"
paths:
  /tweets:
    post:
      consumes:
      - application/json
      description: Create a new tweet for a user
      parameters:
      - description: Tweet object
        in: body
        name: tweet
        required: true
        schema:
          $ref: '#/definitions/http.CreateTweetRequest'
      - description: ID of the current user
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.Tweet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create a new tweet
      tags:
      - tweets
  /tweets/following:
    get:
      consumes:
      - application/json
      description: Get tweets from a list of user IDs with pagination
      parameters:
      - collectionFormat: csv
        description: List of user IDs
        in: query
        items:
          type: string
        name: user_ids
        required: true
        type: array
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.Tweet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get tweets by user IDs
      tags:
      - tweets
schemes:
- http
swagger: "2.0"